Add basic module support
This commit adds basic module support. Modules can be imported from
the current directory or from the standard library.

To be able to refer to things in imported modules, basic expression
parsing is added.

The interpreter now has separate scopes for each module, and has been
refactored to make use of the new Expression parsing.

This also reworked the parser, and moved the parsing logic for each
AST node type into the corresponding AST node class.

The compiler is left behind at this point, and focus is now entirely
on making the interpreter capable enough to start writing the compiler
from scratch in Kod itself.